remove AuthenticationFacade.java and move its logic to an AuthenticationService
code correction according to comments
