+++ src/main/java/online/talkandtravel/service/AuthenticationService.java
@@ -3,16 +3,19 @@ package online.talkandtravel.service;
 import java.io.IOException;
 import online.talkandtravel.model.dto.AuthResponse;
 import online.talkandtravel.model.entity.User;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 /**
  * Service interface for handling user authentication and registration operations.
  *
  * <p>This service is responsible for user registration and login processes, providing mechanisms to
- * create new user accounts and authenticate existing users.
+ * create new user accounts and authenticate existing users, retrieving authenticated user from
+ * spring security context.
  *
  * <p>Methods:
  *
  * <ul>
+ *    <li>{@link #getAuthenticatedUser} gets the authenticated user
  *   <li>{@link #register(User)} - Registers a new user with the provided {@link User} details. This
  *       method handles user creation, including password encryption and any necessary setup. It
  *       returns an {@link AuthResponse} containing information about the authentication result.
@@ -20,11 +23,12 @@ import online.talkandtravel.model.entity.User;
  *       the credentials are valid, it returns an {@link AuthResponse} containing authentication
  *       details.
  * </ul>
- *
  */
 public interface AuthenticationService {
 
   AuthResponse register(User user) throws IOException;
 
   AuthResponse login(String email, String password);
+
+  User getAuthenticatedUser();
 }
